requirements.txt
fastapi
httpx
uvicorn
csscompressor
jsmin
Pillow
pillow-avif-plugin